Relocation bcs_k8s.dashboard package
